agent.core - AOP for TypeScript
