+++ src/lib/agent/core/attribute.ts
@@ -22,40 +22,18 @@ export interface IAttribute {
   
 }
 
-/**
- * Attribute base class
- */
-export class Attribute implements IAttribute {
-  
-  public beforeDecorate(target: Object|Function, targetKey?: string|symbol, descriptor?: PropertyDescriptor): boolean {
-    // always allow decoration by default, unless this method been overwrite by inherited class
-    return true
-  }
-  
-  public getType(): string {
-    throw new TypeError('Not Implemented')
-  }
-  
-  public getInterceptor(): IInterceptor {
-    throw new TypeError('Not Implemented')
-  }
-  
-  generateClassDecorator() {
-    return (target: Function): void => {
-      if (this.beforeDecorate(target)) {
-        const reflection = Reflection.getOwnInstance(target);
-        reflection.addAttribute(this);
-      }
-    };
-  }
-  
-  generateClassMemberDecorator() {
-    return (target: Object, propertyKey: string | symbol, descriptor?: PropertyDescriptor): void => {
-      if (this.beforeDecorate(target, propertyKey, descriptor)) {
-        const reflection = Reflection.getOwnInstance(target, propertyKey, descriptor);
-        reflection.addAttribute(this);
-      }
+export function decorateClass(attribute: IAttribute) {
+  return (target: Function): void => {
+    if (attribute.beforeDecorate(target)) {
+      Reflection.addAttribute(attribute, target);
     }
   }
-  
+}
+
+export function decorateClassMembers(attribute: IAttribute) {
+  return (target: Object, propertyKey: string | symbol, descriptor?: PropertyDescriptor): void => {
+    if (attribute.beforeDecorate(target, propertyKey, descriptor)) {
+      Reflection.addAttribute(attribute, target, propertyKey, descriptor)
+    }
+  }
 }
