+++ src/lib/agent/core/index.ts
@@ -1,0 +1,3 @@
+export { IAttribute, decorateClass, decorateClassMembers } from './attribute'
+export { IInterceptor } from './interceptor'
+export { IInvocation } from './invocation'
