+++ src/lib/agent/core/utils.ts
