+++ src/lib/agent/index.ts
@@ -3,4 +3,4 @@ export * from './core/interceptor'
 export * from './core/invocation'
 export * from './agent'
 export * from './domain'
-export * from './utils'
+export * from './core/utils'
